Add SqlTriggerBindingIntegrationTestBase
